requirements.txt
paho-mqtt
sqlalchemy
pytest
psycopg2-binary
